backend/requirements.txt
asgiref==3.11.1
astroid==4.0.4
attrs==26.1.0
dill==0.4.1
Django==5.2.12
django-cors-headers==4.9.0
django-stubs==6.0.2
django-stubs-ext==6.0.2
djangorestframework==3.17.1
djangorestframework-stubs==3.16.9
drf-spectacular==0.29.0
inflection==0.5.1
isort==8.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
platformdirs==4.9.4
psycopg2-binary==2.9.11
prometheus-client==0.23.1
pylint==4.0.5
pylint-django==2.7.0
pylint-plugin-utils==0.9.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tomli==2.4.1
tomlkit==0.14.0
types-PyYAML==6.0.12.20250915
types-requests==2.33.0.20260402
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
